streamlit_ver/requirements.txt
Pillow==12.1.1
pyserial==3.5
streamlit==1.54.0
